test(parity): complete property-based test suite and Tier 1 CI gate
- Add test_all_pass_certification.py for Property 11 (all-pass verdict aggregation)
- Add test_downstream_blocking.py for Property 9 (downstream stage blocking)
- Add test_first_failing_field.py for Property 10 (first-failing-field identification)
- Update test_prove_exact_sequence_order.py, test_certification_report_fields.py, and oracle loader tests with refined assertions
- Update tasks.md to mark Properties 8–11 complete and Tier 1 CI gate checkpoint verified
- Refactor energy resumability, vertex painting, and math utilities to support exact sequence certification
- Fix edge cleanup vertex degree handling and MATLAB linspace bit-exactness for upsample mesh
- Update parity CLI to surface first-failing-field diagnostics and certification verdicts
- Mark Tier 1 regression gate passing; Tier 2 crop harness ready for next phase

+++ slavv_python/utils/math_utils.py
@@ -47,8 +47,8 @@ def slavv_round(x: float | np.floating) -> int:
     """
     val = float(x)
     if val >= 0.0:
-        return int(math.floor(val + 0.5))
-    return int(math.ceil(val - 0.5))
+        return math.floor(val + 0.5)
+    return math.ceil(val - 0.5)
 
 
 def slavv_round_array(x: np.ndarray) -> np.ndarray:
